editor (feature): wvlet ui -w (workdir) option
